♻️ refactor: drawingTool マシンのコード統合とリネーム
- レガシーのdrawingTool.tsとcreateDrawingTool.tsを統合
- createDrawingTool.tsをdrawingTool.tsにリネーム
- 不要な重複コードを削除してコードベースをシンプル化
- 全てのインポートパスを更新
